refactor: sizes are int64, not uint64
File sizes come from os.FileInfo.Size() and are stored in SQLite, both
int64, yet FTypeStat.NumBytes and the size parameter of
UpdateFileStats/UpdateFileStatsMulti were uint64. Every size made an
unchecked int64 -> uint64 -> int64 round trip, which gosec reports as
G115. Sizes are now int64 end to end, and none of those conversions are
left.

int64 also suits sizes better, because a difference of sizes can be
negative instead of wrapping around.

gogenutils.ByteCountSI takes a uint64, so it is replaced by
types.ByteCountSI(int64). The output format is the same, and a negative
value renders with a minus sign.

API change for consumers: FTypeStat.NumBytes, and the size parameters of
UpdateFileStats and UpdateFileStatsMulti, are now int64.

Closes #6

+++ filetypestats_legacy.go
@@ -61,7 +61,7 @@ func fileTypeStatsDB(scanRoot string, fdb *ftsdb.FileTypeStatsDB) error {
 				fi, err = os.Stat(osPathname)
 				if err == nil {
 					if ftype, err = filetype.FileClass(osPathname); err == nil {
-						return fdb.UpdateFileStatsMulti(osPathname, ftype, uint64(fi.Size()), piBuf)
+						return fdb.UpdateFileStatsMulti(osPathname, ftype, fi.Size(), piBuf)
 					}
 				}
 			}
